tests - rework native-environment.mk
This appears to be something broken with the combination of
windows and git-bash and Python, unrelated to pymake, so I
don't feel too guilty about "cheating" by reworking the test
implementation to avoid writing any newlines from python to
disk and reading them in again in order.

More detail:

Something is not right with round tripping a string containing
a windows-style line ending ("\r\n") to file and back again
using Python3 inside Travis CI Windows' shell environment. In
particular, if we run the following script outside of pymake:

```
with open('a.txt', 'w') as f:
    f.write('hello\r\n')
with open('a.txt', 'r') as f:
    y = f.read()
assert (y == 'hello\n' or y == 'hello\r\n'), y
```

we observe that the string `y` has the value `'hello\n\n'`.
That is, round tripping the string to file and back to a string
has not conserved the number of line endings!

Travis' reference documentation for the windows build environment
states that

  > Git BASH is the shell that’s used to run your build.
  > See Git for Windows for more details.

https://docs.travis-ci.com/user/reference/windows/#git-bash